Add a read-only interface and restrict the reads
This read-only interface is basically just a callback to
upstream environment `get`, and behaves much the same way that
read only environments currently behave in ocaml. The produce_value
function is now a static method that takes a read-only interface
as input, so it has no access to the parent read/write.

This is how I also want the read-only *tables* to behave in an
overlayable refactor of ocaml.

I also made a few other changes that wind up making the situation
more like ocaml in some ways: the base environment cache is now
global (a class attribute) which is true of *all* caches in
the pyre ocaml codebase today.