Add read-only-based overlay
This change captures the essence of my plan for how to refactor
the ocaml env stack and allow overlays:
- All read-only data access happens via function calls (that change
  was made in `read_only`
- As a result, we're able to implement `overlay` by wrapping function
  calls to the parent environment with logic to "intercept" lookups
  that are in the overlay scope.

This implementation uses no caching in the overlay, but it's hopefully
clear that inside the definition of `get` we could close over a dict
and use that as a cache if we wanted; this is very similar to the
ocaml situation where our overlay can close over the definition of
some cache table if we need.

Note that I don't pass dependencies in `get` functions. That's an
optimization that is mostly specific to the python implementation,
since we don't handle failed lookups in this toy. The real ocaml
code is robust to a key not existing (everythign returns optional
values) so it shouldn't matter much whether or not we include
dependencies, we can just do whichever is easiest given the actual
code structure.