writable: factor out the "global" memory in CodeEnv.
Problem: The main hindrance towards refactoring the environment stack
seems to be that CodeEnv directly depends on a global variable. Steven
deliberately added this to mimic Pyre's non-first-class global shared
memory table.

As a first step towards making this more amenable to refactoring, I'm
factoring out that global table. Instead of having `CodeEnv` refer
directly to `CodeEnv.codes`, I'm having it depend on `WritableCodeEnv`.

In this commit, it directly uses `WritableCodeEnv.codes`, so there's no
change in behavior. It's still a naughty singleton.

Also note that I'm making `WritableEnv` accept a `cached` argument so
that `CodeEnv` will pass `cached=WritableCodeEnv.get_env()`. This way,
the only place in `CodeEnv` that knows about `WritableCodeEnv` is the
constructor.