writable: make `produce_value` take a `current_env_getter`.
Problem: We wanted to make `produce_value` a `staticmethod` so that it
didn't depend on the writable memory of the current environment. It can
only depend on the given key and the upstream environment.

But, if `produce_value` is a staticmethod, then the upstreamest
environment - `CodeEnv` - is forced to use a global variable. It does
not have any other place from which to get the code for the given
module. This tightly couples the `CodeEnv` class to the implementation
of the shared-memory hash table and makes it impossible to change the
implementation without changing the environment code. We need to be
programming to an interface, not an implementation.

This commit passes in an optional getter with read-only access to the
current environment memory.