writable: make `WritableCodeEnv` a constructor arg.
Problem: `CodeEnv` still knows about the singleton `WritableCodeEnv`.

Let's factor it out by passing it in the constructor. This way, only the
initialization code knows about `WritableCodeEnv` (or, for that matter,
the codes dictionary used to build the code environment). `CodeEnv` is
agnostic to the actual implementation of its writable environment, which
is what we wanted.