wrap_memory: add WrappedCacheTable to wrap a cache table.
This allows us to intercept reads and writes to the original. For
unsaved files, we read from the wrapper table for that file but defer to
the wrapped table for other files. When there's an update, we only write
those corresponding to the unsaved module; we drop writes made to other
modules.