two_tables: maintain two tables corresponding to saved and unsaved DAGs.
There are two DAGs - one for the saved contents and one for the unsaved contents. We need to be careful to separate them.

Problem: Each operation has two variants: one that works on the original DAG and the other that works on the new DAG.

For example, suppose computing a value for an unsaved module `b` depends
on a saved module class `a.X`, but `a.X` is uncached. In this case, we
have to compute the value for `a.X` ... but in the original DAG. This is
because we do not want `a.X` to use the unsaved values of any of its
dependents. So, we may need to compute `b.XParent` - the parent of `a.X`
- using the saved contents of `b`, even though `b` is unsaved.

Solution: Instead of wrapping the original cache table, as I was doing earlier, we will now just keep two cache tables. The first one is the big one, containing all saved modules and their values. The second one just contains unsaved modules.

There is no need to wrap the cache table or do any pass-through. We will
automatically look up or write to the correct table based on what the
request is - whether the `get` call expects you to use the saved or
unsaved values of your dependents and whether the `update` push call is
for a saved file or an unsaved file.