Make an env-wrapping implementation
It is considerably cleaner to wrap a parent environment
rather than implement the overlay all in a single layer,
because this makes it more or less impossible for the parent
to accidentally mutate due to changes in an overlay (the current
code is messy and it's still possible, but much harder, to create
such a bug; cleaner code would make this easier).

This requires having a first-class read-write cache rather than
a hardcoded global cache as in wrap_memory.py

The current plan is that we can probably make the caches first
class by adding an overlay id to the hash key, which will allow
various overlays to coexist in the same shared memory hash table
without being able to collide with one another. The implementation
detail can be hidden behind a module interface.