Add overlay_keys.py - a prototype of latest plan
Our latest plan for ocaml is to attempt to make full pushable
overlays and implement them using a single module per shared
memory, by simulating first-class mutable maps using overlay
keys.

This python implementation highlights the idea: if we're forced
to use only one dictionary per environment storage, then we can
create multiple "synthetic" dictionaries by using the identity
of a dictionary as a prefix on our hash key.

In Python this feels absurd since it is so easy to do something
more reasonable, but in ocaml we have a lot of constraints - making
separate shared memory tables per overlay might well do nasty
things to our overall memory profile, so allocating tables globally
makes sense from a runtime point of view, and we only want
first-class values from a logical point of view.

We'll be able to hide the fact that we're doing this using the
ocaml module system, which will also make it easier to migrate
off of a single shared memory table later if we want to.