Add an initial attempt at an environment model
So far it isn't working: my model is over-constrained so that
I cannot generate any transitions. Unfortunately I'm actually not
yet skilled enough with alloy to debug an over-constrained mode;
with under-constrained models the examples are very useful but
with over-contrained ones the off-the-shelf tools don't help much.

Hoping to iterate on this later!